nice-main/stock_etl_project/requirements.txt
requests
pandas
sqlalchemy
psycopg2-binary
clickhouse-driver
lightgbm
matplotlib
scikit-learn
joblib